+++ pt_os_web_portal/tests/.old/test_finalise.py
@@ -2,7 +2,6 @@ from os import path
 from shlex import split
 
 from flask import json
-
 from tests.data.finalise_data import (
     available_space,
     available_space_out,
