+++ pt_os_web_portal/tests/.old/test_locales.py
@@ -1,5 +1,4 @@
 from flask import json
-
 from tests.data.locale_data import default_locale, formatted_locales
