+++ pt_os_web_portal/tests/.old/test_os_updater_class.py
