+++ pt_os_web_portal/tests/.old/test_registration.py
@@ -1,7 +1,6 @@
 from os.path import isfile
 
 from helpers.registration import pi_top_registration_txt
-
 from tests.utils import assert_file_content
