+++ pt_os_web_portal/tests/.old/test_wifi_manager.py
@@ -1,5 +1,4 @@
 from flask import json
-
 from tests.data.wifi_manager_data import wifi_ssids, wpa_cli_status
 from tests.utils import dotdict
