+++ pt_os_web_portal/tests/data/wifi_manager_data.py
